requirements.txt
gpxpy~=1.4.2
pyserial~=3.5
pandas~=1.2.4